requirements.txt
firebase-admin==5.3.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.34
selenium==4.24.0
webdriver-manager==4.0.2
python-dotenv==1.0.1
pytz==2024.1
fastapi==0.114.0
uvicorn==0.30.6
fastapi-utils==0.7.0
typing-inspect==0.9.0
beautifulsoup4==4.12.3
fastapi-cors==0.0.6
alembic-1.13.3